lecture-planner/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
google-generativeai==0.8.5
python-dotenv==1.1.1
streamlit==1.47.1
PyPDF2==3.0.1
requests==2.32.4